feat: GameTabs Keybinds added
This is a complete rewrite of my original keybindings in 1.4.
Much cleaner and simpler than the original, partially due to knowledge, partially due to being built directly on the TRSGameTabs.

The way it works is the same as the original in that we try random FKeys when we try to use it, probably fail, but cache the results for next uses. Once things start getting mapped it starts working as intended.