fix: added a 50ms delay to MainScreen.IsUpText
This is the same as 1.4 version and is what was missing to make up text reliable.
Things that depended on up text that were working very weirdly should now behave much better.
On top of that, this keeps the new design I did for 2.0 in which is does a 85% percentage match instead of comparing the uptext 1 to 1.
Also fixed some bugs in Map jsons