feat(TModel): read notes
TModel is an interace for 3D models. It supports a simple model and material file loader as well as simple TCuboid creation.
It's not in use for anything yet but the idea is that you cuold pick in the future to use the default which is the cuboid everyone is used to or actually load the object/npc model.